support for other index types in TC
