+++ cpp/oneapi/dal/algo/triangle_counting/detail/select_kernel.cpp
@@ -15,6 +15,7 @@
 *******************************************************************************/
 
 #include "oneapi/dal/algo/triangle_counting/detail/select_kernel.hpp"
+#include "oneapi/dal/algo/triangle_counting/detail/vertex_ranking_default_kernel.hpp"
 #include "oneapi/dal/algo/triangle_counting/backend/cpu/vertex_ranking_default_kernel.hpp"
 #include "oneapi/dal/algo/triangle_counting/backend/cpu/relabel_kernels.hpp"
 #include "oneapi/dal/detail/policy.hpp"
@@ -22,7 +23,8 @@
 
 namespace oneapi::dal::preview::triangle_counting::detail {
 
-std::int64_t triangle_counting_global_scalar(const dal::detail::host_policy& policy,
+template<>
+std::int64_t triangle_counting_global_scalar<std::int32_t>(const dal::detail::host_policy& policy,
                                              const std::int32_t* vertex_neighbors,
                                              const std::int64_t* edge_offsets,
                                              const std::int32_t* degrees,
@@ -37,7 +39,8 @@ std::int64_t triangle_counting_global_scalar(const dal::detail::host_policy& pol
     });
 }
 
-std::int64_t triangle_counting_global_vector(const dal::detail::host_policy& policy,
+template<>
+std::int64_t triangle_counting_global_vector<std::int32_t>(const dal::detail::host_policy& policy,
                                              const std::int32_t* vertex_neighbors,
                                              const std::int64_t* edge_offsets,
                                              const std::int32_t* degrees,
@@ -52,7 +55,8 @@ std::int64_t triangle_counting_global_vector(const dal::detail::host_policy& pol
     });
 }
 
-std::int64_t triangle_counting_global_vector_relabel(const dal::detail::host_policy& policy,
+template<>
+std::int64_t triangle_counting_global_vector_relabel<std::int32_t>(const dal::detail::host_policy& policy,
                                                      const std::int32_t* vertex_neighbors,
                                                      const std::int64_t* edge_offsets,
                                                      const std::int32_t* degrees,
@@ -67,7 +71,8 @@ std::int64_t triangle_counting_global_vector_relabel(const dal::detail::host_pol
     });
 }
 
-array<std::int64_t> triangle_counting_local(
+template<>
+array<std::int64_t> triangle_counting_local<std::int32_t>(
     const dal::detail::host_policy& policy,
     const dal::preview::detail::topology<std::int32_t>& data,
     int64_t* triangles_local) {
